fix: CFD editor correctness fixes
- Fix bigint IPC crash: i64 → f64 in all FieldValue/DictKey Int/Enum fields
  so JSON serialization works through Tauri invoke
- Populate target_file on Ref values for cross-file navigation
- Router: single state object fixes stale-closure bug in push/replace/back
- TableView: sync activeType when type list changes after reload; accept
  initialTypeFilter prop from route
- useProject: add refreshSnapshot() (re-opens project) and saveNow()
  (flushes dirty debounce immediately)
- App: wire Ctrl+S to saveNow; refresh full snapshot after create_file so
  file tree updates; pass typeFilter from route to TableView
- Housekeeping: gitignore gen/ and dist/ auto-generated dirs, remove them
  from tracking

+++ crates/coflow-editor-core/src/types.rs
@@ -60,10 +60,10 @@ pub struct FieldCell {
 pub enum FieldValue {
     Null,
     Bool { v: bool },
-    Int { v: i64 },
+    Int { v: f64 },
     Float { v: f64 },
     Str { v: String },
-    Enum { enum_name: String, variant: String, int_value: i64 },
+    Enum { enum_name: String, variant: String, int_value: f64 },
     Object { actual_type: String, fields: Vec<FieldCell> },
     Ref { target_type: String, target_key: String, target_file: Option<String> },
     Array { items: Vec<FieldValue> },
@@ -82,8 +82,8 @@ pub struct DictEntry {
 #[ts(export)]
 pub enum DictKey {
     Str { v: String },
-    Int { v: i64 },
-    Enum { enum_name: String, variant: String, int_value: i64 },
+    Int { v: f64 },
+    Enum { enum_name: String, variant: String, int_value: f64 },
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, TS)]
